Move score recalculator to team viewset
