+++ src/core/urls.py
@@ -29,7 +29,6 @@ urlpatterns = [
     path("hints/", include("hint.urls")),
     path("leaderboard/", include("leaderboard.urls")),
     path("member/", include("member.urls")),
-    path("scorerecalculator/", include("scorerecalculator.urls")),
     path("stats/", include("stats.urls")),
     path("team/", include("team.urls")),
     path("pages/", include("pages.urls")),
